+++ src/Viserio/Component/Routing/AbstractRouteDispatcher.php
@@ -255,99 +255,6 @@ abstract class AbstractRouteDispatcher
     }
 
     /**
-     * Run the given route within a Stack "onion" instance.
-     *
-     * @param \Viserio\Component\Contracts\Routing\Route $route
-     * @param \Psr\Http\Message\ServerRequestInterface   $request
-     *
-     * @return \Psr\Http\Message\ResponseInterface
-     */
-    protected function runRouteWithinStack(RouteContract $route, ServerRequestInterface $request): ResponseInterface
-    {
-        $middlewares = $this->gatherRouteMiddleware($route);
-
-        return (new Pipeline())
-            ->setContainer($this->getContainer())
-            ->send($request)
-            ->through($middlewares)
-            ->then(function ($request) use ($route) {
-                return $route->run($request);
-            });
-    }
-
-    /**
-     * Gather the middleware for the given route.
-     *
-     * @param \Viserio\Component\Contracts\Routing\Route $route
-     *
-     * @return array
-     */
-    protected function gatherRouteMiddleware(RouteContract $route): array
-    {
-        $middlewares      = [];
-
-        self::map($route->gatherMiddleware(), function ($name) use (&$middlewares, $route) {
-            $middlewares[] = MiddlewareNameResolver::resolve(
-                $name,
-                $this->middlewares,
-                $this->middlewareGroups,
-                $route->gatherDisabledMiddlewares()
-            );
-        });
-
-        return (new SortedMiddleware(
-            $this->middlewarePriority,
-            array_values(self::flatten($middlewares))
-        ))->getAll();
-    }
-
-    /**
-     * Applies the callback to the elements of the given arrays.
-     *
-     * @param array    $array
-     * @param callable $callback
-     *
-     * @return array
-     */
-    protected static function map(array $array, callable $callback): array
-    {
-        $newArray = [];
-
-        foreach ($array as $key => $item) {
-            $result = $callback($item, $key);
-
-            $newArray = is_array($result) ?
-                array_replace_recursive($array, $result) :
-                array_merge_recursive($array, (array) $result);
-        }
-
-        return $newArray;
-    }
-
-    /**
-     * Flatten a nested array to a separated key.
-     *
-     * @param array  $array
-     * @param string $prepend
-     *
-     * @return array
-     */
-    protected static function flatten(array $array, string $prepend = ''): array
-    {
-        $flattened = [];
-
-        foreach ($array as $key => $value) {
-            if (is_array($value)) {
-                $flattened = array_merge($flattened, static::flatten($value, $prepend . $key));
-            } else {
-                $flattened[$prepend . $key] = $value;
-            }
-        }
-
-        return $flattened;
-    }
-
-    /**
      * Make a nested path, creating directories down the path recursion.
      *
      * @param string $path
